fix: Add Suspense boundary to chat page for useSearchParams
- Refactored chat page to use Suspense boundary as required by Next.js
- Extracted chat logic to separate ChatClient component
- Fixes Vercel deployment error with useSearchParams
